+++ pluralize/index.d.ts
@@ -1,13 +1,20 @@
 export type pluralizer = (n: number) => string
 
+export const plural_AR: pluralizer
 export const plural_CS: pluralizer
 export const plural_DA: pluralizer
 export const plural_DE: pluralizer
 export const plural_EN: pluralizer
 export const plural_ES: pluralizer
 export const plural_FR: pluralizer
+export const plural_FI: pluralizer
+export const plural_FO: pluralizer
 export const plural_IS: pluralizer
 export const plural_IT: pluralizer
 export const plural_JA: pluralizer
+export const plural_NB: pluralizer
+export const plural_NO: pluralizer
 export const plural_PT: pluralizer
+export const plural_PL: pluralizer
 export const plural_SE: pluralizer
+export const plural_TH: pluralizer